Chore: chart Reducer를 chart-list Reducer로 rename
