+++ src/reducer/chart-list.js
